ai_service/.history/requirements_20250815200115.txt
# Core libraries
pandas>=2.0.0
numpy>=1.24.0

# Vector database
chromadb>=0.4.0

# Optional: for embeddings if using sentence-transformers
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# Progress bar (useful during data processing)
tqdm>=4.65.0

# Web UI
gradio>=3.40.0

google-generativeai

